Abstract monster and wall classes
